ai_powered_personal_assistant/requirements.txt
flask
openai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client